backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12
openai>=1.30.0
python-dotenv>=1.0.0
